+++ src/functions/index.ts
@@ -1,2 +1,1 @@
-export * from "./poolList";
 export * from "./ReservesList";
