+++ src/models/Pools.ts
@@ -1,0 +1,4 @@
+export interface PoolViewModel {
+    name?: string;
+    address: string;
+}
